refactor(tui): extract sessions modal and session lifecycle component
